teste_3/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
greenlet==2.0.0.post0
idna==3.4
numpy==1.23.4
pandas==1.5.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
tabulate==0.9.0
urllib3==1.26.12
